Move dummy data generation into collection, stub out sync to provide dummies.
