+++ clientapp/templates/landing.html
@@ -3,8 +3,12 @@
     <p>
       Stumptown fugiat street art 8-bit Cosby sweater mumblecore. Intelligentsia slow-carb Austin, lo-fi Pitchfork blog freegan est iPhone.
     </p>
-    <a href="#">Sign Up Now</a>
-    <a href="#" class="button">Log In</a>
+    {% if loggedIn %}
+      <a href="#" class="button">Take the Quiz</a>
+    {% else %}
+      <a href="#">Sign Up Now</a>
+      <a href="#" class="button">Log In</a>
+    {% endif %}
   </div>
   <div class="banner">
     <div class="row text-center">
