+++ clientapp/views/landing.js
@@ -1,20 +1,6 @@
 var HumanView = require('human-view');
 var ListingView = require('./listing');
-var Achievements = require('../models/achievements');
 var templates = require('templates');
-
-function fakeAchievement () {
-  var type = Math.random() < 0.5 ? 'Badge' : 'Pathway';
-  var data = {
-    type: type.toLowerCase(),
-    title: 'A Very Long ' + type + ' Title',
-    tags: ['service', 'barista', 'coffeelover', 'fake'],
-    creator: 'Starbucks'
-  };
-  if (me.loggedIn)
-    data.favorite = Math.random() < 0.2 ? true : false;
-  return data;
-}
 
 module.exports = HumanView.extend({
   template: templates.landing,
@@ -22,9 +8,9 @@ module.exports = HumanView.extend({
     'click a': 'handleLink'
   },
   render: function () {
-    this.renderAndBind({me: this.model});
+    this.renderAndBind(this.model);
     this.renderSubview(new ListingView({
-      collection: new Achievements([1, 2, 3, 4, 5, 6, 7, 8].map(fakeAchievement))
+      collection: this.collection
     }), '.content');
     return this;
   },
